+++ R/compare_parameters.R
@@ -26,6 +26,7 @@
 #'   }
 #' @inheritParams model_parameters.default
 #' @inheritParams model_parameters.cpglmm
+#' @inheritParams print.parameters_model
 #'
 #' @note This function is in an early stage and does not yet cope with more
 #'   complex models, and probably does not yet properly render all model
@@ -34,23 +35,23 @@
 #' @return A data frame of indices related to the model's parameters.
 #'
 #' @examples
-#' if (packageVersion("insight") >= "0.13.0") {
-#'   data(iris)
-#'   lm1 <- lm(Sepal.Length ~ Species, data = iris)
-#'   lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
-#'   lm3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
-#'   compare_parameters(lm1, lm2, lm3)
+#' data(iris)
+#' lm1 <- lm(Sepal.Length ~ Species, data = iris)
+#' lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
+#' lm3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
+#' compare_parameters(lm1, lm2, lm3)
 #'
-#'   data(mtcars)
-#'   m1 <- lm(mpg ~ wt, data = mtcars)
-#'   m2 <- glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
-#'   compare_parameters(m1, m2)
+#' data(mtcars)
+#' m1 <- lm(mpg ~ wt, data = mtcars)
+#' m2 <- glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
+#' compare_parameters(m1, m2)
 #'
-#'   # exponentiate coefficients, but not for lm
-#'   compare_parameters(m1, m2, exponentiate = "nongaussian")
+#' \dontrun{
+#' # exponentiate coefficients, but not for lm
+#' compare_parameters(m1, m2, exponentiate = "nongaussian")
 #'
-#'   # change column names
-#'   compare_parameters(m1, m2, column_names = c("linear model", "logistic reg."))
+#' # change column names
+#' compare_parameters(m1, m2, column_names = c("linear model", "logistic reg."))
 #' }
 #' @export
 compare_parameters <- function(...,
@@ -63,6 +64,7 @@ compare_parameters <- function(...,
                                p_adjust = NULL,
                                style = NULL,
                                column_names = NULL,
+                               groups = NULL,
                                verbose = TRUE) {
   models <- list(...)
   model_names <- match.call(expand.dots = FALSE)$`...`