+++ R/methods_aov.R
@@ -41,7 +41,11 @@
 #' @return A data frame of indices related to the model's parameters.
 #'
 #' @note For ANOVA-tables from mixed models (i.e. \code{anova(lmer())}), only
-#'   partial or adjusted effect sizes can be computed.
+#'   partial or adjusted effect sizes can be computed. Note that type 3 ANOVAs
+#'   with interactions involved only give sensible and informative results when
+#'   covariates are mean-centred and factors are coded with orthogonal contrasts
+#'   (such as those produced by \code{contr.sum}, \code{contr.poly}, or
+#'   \code{contr.helmert}, but \emph{not} by the default \code{contr.treatment}).
 #'
 #' @examples
 #' if (requireNamespace("effectsize", quietly = TRUE)) {
@@ -116,9 +120,19 @@ model_parameters.aov <- function(model,
     }
   }
 
+  # try to extract type of anova table
+  if (is.null(type)) {
+    type <- .anova_type(model)
+  }
+
   # exceptions
   if (.is_levenetest(model)) {
     return(model_parameters.htest(model, ...))
+  }
+
+  # check contrasts
+  if (verbose) {
+    .check_anova_contrasts(model, type)
   }
 
   # extract standard parameters
@@ -259,9 +273,9 @@ model_parameters.afex_aov <- function(model,
     with_df_and_p <- summary(model$Anova)$univariate.tests
     params$`Sum Sq` <- with_df_and_p[-1, 1]
     params$`Error SS` <- with_df_and_p[-1, 3]
-    out <- model_parameters(params, df_error = NULL, type = NULL, ...)
+    out <- .extract_parameters_anova(params, test = NULL)
   } else {
-    out <- model_parameters(model$Anova, df_error = NULL, type = attr(model, "type"), ...)
+    out <- .extract_parameters_anova(model$Anova, test = NULL)
   }
 
   out <- .effectsizes_for_aov(
@@ -275,9 +289,12 @@ model_parameters.afex_aov <- function(model,
     ...
   )
 
+  # add attributes
+  out <- .add_anova_attributes(out, model, ci, test = NULL, ...)
+
   # filter parameters
   if (!is.null(parameters)) {
-    params <- .filter_parameters(params, parameters, verbose = verbose)
+    out <- .filter_parameters(out, parameters, verbose = verbose)
   }
 
   if (!"Method" %in% names(out)) {
@@ -285,6 +302,7 @@ model_parameters.afex_aov <- function(model,
   }
 
   attr(out, "title") <- unique(out$Method)
+  class(out) <- unique(c("parameters_model", "see_parameters_model", class(out)))
 
   out
 }
@@ -293,6 +311,105 @@ model_parameters.afex_aov <- function(model,
 
 
 # helper ------------------------------
+
+
+.anova_type <- function(model, type = NULL) {
+  if (is.null(type)) {
+
+    type_to_numeric <- function(type) {
+      if (is.numeric(type)) {
+        return(type)
+      }
+      switch(
+        type,
+        "1" = ,
+        "I" = 1,
+        "2" = ,
+        "II" = 2,
+        "3" = ,
+        "III" = 3,
+        1
+      )
+    }
+
+    # default to 1
+    type <- 1
+
+    if (!is.null(attr(model, "type", exact = TRUE))) {
+      type <- type_to_numeric(attr(model, "type", exact = TRUE))
+    } else if (!is.null(attr(model, "heading"))) {
+      heading <- attr(model, "heading")[1]
+      if (grepl("(.*)Type (.*) Wald(.*)", heading)) {
+        type <- type_to_numeric(trimws(gsub("(.*)Type (.*) Wald(.*)", "\\2", heading)))
+      } else if (grepl("Type (.*) Analysis(.*)", heading)) {
+        type <- type_to_numeric(trimws(gsub("Type (.*) Analysis(.*)", "\\1", heading)))
+      } else if (grepl("(.*)Type (.*) tests(.*)", heading)) {
+        type <- type_to_numeric(trimws(gsub("(.*)Type (.*) tests(.*)", "\\2", heading)))
+      }
+    } else if ("type" %in% names(model) && !is.null(model$type)) {
+      type <- type_to_numeric(model$type)
+    }
+  }
+
+  type
+}
+
+
+.check_anova_contrasts <- function(model, type) {
+  # check only valid for anova tables of type III
+  if (!is.null(type) && type == 3) {
+
+    # check for interaction terms
+    interaction_terms <- tryCatch(
+      {
+        insight::find_interactions(model, flatten = TRUE)
+      },
+      error = function(e) {
+        if (is.data.frame(model)) {
+          if (any(grepl(":", row.names(model), fixed = TRUE))) {
+            TRUE
+          } else {
+            NULL
+          }
+        }
+      }
+    )
+
+    # try to access data of model predictors
+    predictors <- tryCatch(
+      {
+        insight::get_predictors(model)
+      },
+      error = function(e) {
+        NULL
+      }
+    )
+
+    # if data available, check contrasts and mean centering
+    if (!is.null(predictors)) {
+      treatment_contrasts_or_not_centered <- sapply(predictors, function(i) {
+        if (is.factor(i)) {
+          cn <- stats::contrasts(i)
+          if (is.null(cn) || (all(cn %in% c(0, 1)))) {
+            return(TRUE)
+          }
+        } else {
+          if (abs(mean(i, na.rm = TRUE)) > 1e-2) {
+            return(TRUE)
+          }
+        }
+        return(FALSE)
+      })
+    } else {
+      treatment_contrasts_or_not_centered <- FALSE
+    }
+
+    # successfully checked predictors, or if not possible, at least found interactions?
+    if (!is.null(interaction_terms) && (any(treatment_contrasts_or_not_centered) || is.null(predictors))) {
+      message(insight::format_message("Type 3 ANOVAs only give sensible and informative results when covariates are mean-centered and factors are coded with orthogonal contrasts (such as those produced by 'contr.sum', 'contr.poly', or 'contr.helmert', but *not* by the default 'contr.treatment')."))
+    }
+  }
+}
 
 
 .effectsizes_for_aov <- function(model,