+++ tests/testthat/test-MCMCglmm.R
@@ -39,7 +39,15 @@ if (require("testthat") &&
 
   test_that("model_parameters", {
     expect_equal(
-      model_parameters(m1)$Median,
+      model_parameters(m1, centrality = "mean")$Mean,
+      c(1.0132, 0.04232),
+      tolerance = 0.01
+    )
+  })
+
+  test_that("model_parameters", {
+    expect_equal(
+      model_parameters(m1, centrality = "median")$Median,
       c(1.01382, 0.04207),
       tolerance = 0.01
     )