+++ tests/testthat/test-model_parameters.BFBayesFactor.R
@@ -20,7 +20,7 @@ if (require("testthat") &&
   test_that("model_parameters.BFBayesFactor", {
     set.seed(123)
     model <- BayesFactor::anovaBF(Sepal.Length ~ Species, data = iris)
-    expect_equal(parameters::model_parameters(model)$Median, c(5.8431, -0.8266, 0.092, 0.734, 0.2681, 2.0415), tolerance = 2)
+    expect_equal(parameters::model_parameters(model, centrality = "median")$Median, c(5.8431, -0.8266, 0.092, 0.734, 0.2681, 2.0415), tolerance = 2)
   })
 
   df <- mtcars
@@ -37,18 +37,18 @@ if (require("testthat") &&
   test_that("model_parameters.BFBayesFactor", {
     set.seed(123)
     model <- BayesFactor::anovaBF(mpg ~ gear * am, data = df)
-    expect_equal(model_parameters(model)$Median, c(20.660577, -3.244485, 3.244485, 25.128301, 0.778598, NA, NA, NA), tolerance = 1L)
+    expect_equal(model_parameters(model, centrality = "mean")$Mean, c(20.7099, -3.24884, 3.24884, 26.51413, 5.30506, NA, NA, NA), tolerance = 1L)
   })
 
   if (.runThisTest) {
     data(raceDolls)
     bf <- contingencyTableBF(raceDolls, sampleType = "indepMulti", fixedMargin = "cols")
-    mp <- suppressWarnings(model_parameters(bf, verbose = FALSE))
+    mp <- suppressWarnings(model_parameters(bf, centrality = "mean", dispersion = TRUE, verbose = FALSE))
 
     test_that("model_parameters.BFBayesFactor", {
       expect_equal(colnames(mp), c(
-        "Parameter", "Median", "CI", "CI_low", "CI_high", "pd", "ROPE_Percentage",
-        "Prior_Distribution", "Prior_Location", "Prior_Scale", "BF", "Method"
+        "Parameter", "Mean", "CI", "CI_low", "CI_high", "pd", "ROPE_Percentage",
+        "SD", "Prior_Distribution", "Prior_Location", "Prior_Scale", "BF", "Method"
       ))
     })
 
@@ -57,11 +57,11 @@ if (require("testthat") &&
       data = puzzles, whichRandom = "ID",
       whichModels = "top", progress = FALSE
     )
-    mp <- model_parameters(result, verbose = FALSE)
+    mp <- model_parameters(result, centrality = "median", dispersion = TRUE, verbose = FALSE)
 
     test_that("model_parameters.BFBayesFactor", {
       expect_equal(colnames(mp), c(
-        "Parameter", "Median", "CI", "CI_low", "CI_high", "pd", "ROPE_Percentage",
+        "Parameter", "Median", "MAD", "CI", "CI_low", "CI_high", "pd", "ROPE_Percentage",
         "Prior_Distribution", "Prior_Location", "Prior_Scale", "Effects",
         "Component", "BF", "Method"
       ))