+++ tests/testthat/test-model_parameters.mixed.R
@@ -137,23 +137,21 @@ if (.runThisTest &&
       )
     )
 
-    if (packageVersion("insight") > "0.13.2") {
-      out <- utils::capture.output(print(model_parameters(m1, effects = "fixed", summary = TRUE)))
-      expect_equal(
-        out,
-        c(
-          "# Fixed Effects",
-          "",
-          "Parameter   | Coefficient |   SE |        95% CI | t(28) |      p",
-          "-----------------------------------------------------------------",
-          "(Intercept) |        0.65 | 0.50 | [-0.34, 1.64] |  1.29 | 0.196 ",
-          "cyl         |        0.40 | 0.08 | [ 0.25, 0.55] |  5.29 | < .001",
-          "",
-          "Model: wt ~ cyl (32 Observations)",
-          "Residual standard deviation: 0.594 (df = 28)",
-          "Conditional R2: 0.628; Marginal R2: 0.550"
-        )
+    out <- utils::capture.output(print(model_parameters(m1, effects = "fixed", summary = TRUE)))
+    expect_equal(
+      out,
+      c(
+        "# Fixed Effects",
+        "",
+        "Parameter   | Coefficient |   SE |        95% CI | t(28) |      p",
+        "-----------------------------------------------------------------",
+        "(Intercept) |        0.65 | 0.50 | [-0.34, 1.64] |  1.29 | 0.196 ",
+        "cyl         |        0.40 | 0.08 | [ 0.25, 0.55] |  5.29 | < .001",
+        "",
+        "Model: wt ~ cyl (32 Observations)",
+        "Residual standard deviation: 0.594 (df = 28)",
+        "Conditional R2: 0.628; Marginal R2: 0.550"
       )
-    }
+    )
   })
 }